feat(BookingMicroservice): Implementing the put and delete methods in the service.
Also adding a new endpoint in the room service to validate when a room is occupied at the time when a booking is being updating. Has also been added new exceptions in the booking service. Moreover, the feign clients also has being refactored and expanded based on the new logic of the put and delete methods.