docs(RoomMicroservice): Adding swagger to controllers and DTOs
why: The service didn't have implemented swagger and the documentation was poor
what: Implement swagger annotations to document the methods of all controllers and the controllers itself. Also the DTOs and the response object was documented to enhance the comprehension of the architecure and logic.
impact: Big improvement in the comprehension of the architecture and code logic. Besides, it strengthen the project and make it more robust because now the service has documented each endpoint and schemas.