docs(General): Updating README file with actual changes
why: Because the README of the project is empty and is necessary to enhance the project, not only the code, also the documentation.
what: It was updated with the architecture style, diagrams with mermaid, a brief introduction, tech stack, explanation of each microservice, documentation with swagger and also an explanation of how to run locally.
impact: Improve the understanding for people that visits the project.